bugfix: functional beamforming works for all steer types + structural changes to beamformers
Previously the functional beamformer did only work well for the steering formulations I and IV. This was due to the fact that the steering vectors were not properly normalized. In order to reach that goal, the beamformers were changed: Now all predefined beamformers return a vector that contains the value that the beamformer output needs to be multilpied with, in order to fullfill: steer^H * steer=1.  Furthermore the functionalBeamformer is changed in order to work correctly for all steering vector types. It now also works better if the diagonal is removed--> the root is directly apllied to the csm (doesnt use eigenValues as the eigenVectors wouldnt be orthogonal anymore). Doesnt work very well!  The Validation script is updated to the new structure as well.