feature: user-defined precision of csm, beamformer and psfs
The user can now specify the precision of the CSM (influences eigenvalue decomp as well).

Furthermore the user can do the same for the beamformer results. Here only the storage is variable -> the calculation is always done in float64

Lastly the user can define the precision of the psf for the methods damas, damasPlus and clean-psf.
The precision of the results is in those case independent of the precision of the psfs.