Introduce SteeringVector-class
Adjustments to the general structure of the beamformers: All the mic, grid, environment, speed of sound, ..., information is now stored in a steer_obj property which is an instance of the new SteeringVector class.
Subfunctions of this class controll the calling of beamformer/PSF/transfer core-funtionalities.
Due to this process clean-sc looks a little prettier now.